+++ lib/matheus.rb
@@ -9,4 +9,5 @@ Dotenv.load("~/.env")
 Zeitwerk::Loader.for_gem.setup
 
 module Matheus
+  QUESTION_HISTORY_FILE = File.expand_path("~/.qa_history.json")
 end