+++ lib/matheus/q.rb
@@ -1,5 +1,6 @@
 require "openai"
 require "tty-markdown"
+require "json"
 
 module Matheus
   # Usage:
@@ -9,7 +10,10 @@ module Matheus
     BASE_PROMPT = "Answer this question in a short and concise way. You can use markdown in the response: "
 
     def call(question)
-      ask_llm(question).then { print_markdown _1 }
+      question = question.join(" ")
+      answer = ask_llm(question)
+      print_markdown(answer)
+      save_qa(question, answer)
     rescue => e
       Failure(e.message)
     end
@@ -39,6 +43,16 @@ module Matheus
 
     def client
       OpenAI::Client.new(access_token: ENV.fetch("OPENAI_API_KEY"))
+    end
+
+    def save_qa(question, answer)
+      history = load_history
+      history << { question:, answer:, timestamp: Time.now.to_s }
+      File.write(QUESTION_HISTORY_FILE, JSON.pretty_generate(history))
+    end
+
+    def load_history
+      File.exist?(QUESTION_HISTORY_FILE) ? JSON.parse(File.read(QUESTION_HISTORY_FILE)) : []
     end
   end
 end